requirements.txt
streamlit
torch
opencv-python
numpy
Pillow